@Sean-Der 2022-05-21
That's really exciting! @adriancable how/why it is faster? I don't understand what this is doing fully

@jech Do you have any concerns about this/see anything that could go wrong?