@adriancable 2022-05-21
It’s definitely worth getting extra eyes on, but the performance gain is just so huge. I guess I have done a lot of embedded work where everything is unsafe and these kinds of tricks are used all the time, and so in fact when things operate one byte at a time (like it used to) it’s a red flag for me.

I actually think this change is actually conpletely safe and is functionally identical to the original.