@jech 2022-05-21
Good catch, it's definitely an optimisation worth doing.

Adrian's version doesn't handle the case where the buffers are not aligned, which might possibly cause difficult to reproduce issues on arches that don't handle unaligned stores (at least ARMv5, MIPS and Sparc, don't know about RISC-V).  (I'm not saying it will actually cause trouble, I'm just saying that the possibility makes me nervous).

However, it turns out that we already have an optimised version of xorBytes that handles this case correctly, and it's also more efficient than Adrian's version, since it handles 64-bit arches better.  It's here: https://github.com/pion/stun/blob/master/xor.go.  Unfortunately, it doesn't currently do ARMv7, as discussed here: https://github.com/pion/stun/pull/73#issuecomment-767902947.

(Note further that the Go stdlib sources contain a version of xorBytes that's written in assembler, which used 128-bit copies on ARM (using NEON) and 256-bit copies on AMD64 (using AVX).  It's under `src/crypto/cypher`.  Unfortunately, it's not exported.)

So I suggest the following approach:
  - move `xorBytes` from pion/stun into a repository on which both stun and srtp may depend, and use the same function in both places;
  - extend the condition at https://github.com/pion/stun/blob/master/xor.go#L14 to include ARMv6 and ARMv7 (but not ARMv5);
  - consider adding assembler versions from the Go sources if benchmarks show that it is worthwile and Sean agrees.
Ideally, of course, we'd convince the Go people to export their version so we can use it, but I have little hope given how unresponsive they have been in the past.