@jech 2022-05-21
https://github.com/golang/go/issues/53021