@adriancable 2022-05-21
Juliusz - thanks for the great feedback. I looked at the `sec/crypto/cipher` implementation and unfortunately this doesn’t include ARMv7 or ARMv6 when setting `unalignedAccess`. Forget about assembly implementations for a moment - on ARM32 they’re not even enabling the unaligned access optimisation. Do you know why this might be? Is this a golang bug/oversight? If so I will file an issue.