@jech 2022-05-21
No idea.  Will you file an issue, or shall I?