@Sean-Der 2022-05-21
It sounds like this could end up in `/x` that is pretty exciting!

@adriancable What do you think of putting it in https://github.com/pion/transport/ since everyone already depends on that. If you want to start github.com/pion/crypto I am up for that as well. The ops burden is pretty significant with all the repos we have today already.